requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
dj-database-url==2.3.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==23.0.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
python-nmap==0.7.1
pytz==2024.1
railway==0.0.4
requests==2.31.0
setuptools==69.1.1
sqlparse==0.4.4
structlog==24.4.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
waitress==3.0.1
whitenoise==6.8.2
